Коллектор: исправить опрос DID топлива и запись CSV
- добавить DID 0x0014 (curr_fuel_tank) в список циклического опроса,
  чтобы колонка "Период" показывала фактическое значение, приходящее от МК;
- перевести планировщик опроса с жёсткой схемы на 2 параметра на обход
  произвольного списка DID, чтобы цикл корректно работал для 0x0014,
  0x0018 и 0x0019;
- убрать подмену поля period интервалом опроса при обработке DID 0x0018
  и 0x0019, чтобы значение периода больше не затиралось;
- изменить обработку DID 0x0018: принимать уровень топлива как целое
  значение в десятых долях процента (0..1000), конвертировать на стороне
  ПК в диапазон 0.0..100.0 через деление на 10 и ограничивать результат;
- отображать топливо в таблице с точностью до одной десятой процента;
- сохранять fuel_percent в CSV как число с плавающей точкой вместо целого,
  чтобы в файле оставалась точность 0.1 процента;
- обновить README описанием формата DID 0x0018 и тем, что в UI/CSV
  значение отображается после деления на 10;
- восстановить метод AppController._set_programming_active(), который был
  потерян как метод класса, из-за чего при нажатии кнопки "Старт" для
  записи CSV возникал AttributeError.

+++ ui/qml/collector_csv_manager.py
@@ -45,12 +45,12 @@ class CollectorCsvManager:
         measurement_time: str,
         period_ticks: int,
         temperature_c: float,
-        fuel_percent: int,
+        fuel_percent: float,
     ):
         row = (
             self._format_value(str(measurement_time)),
             self._format_value(int(period_ticks)),
             self._format_value(float(temperature_c)),
-            self._format_value(int(fuel_percent)),
+            self._format_value(float(fuel_percent)),
         )
         self._append(self._csv_path, row)
